consolidate gibbs priors
